Refactor to move cafile package
